backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
reportlab==4.0.7
qrcode[pil]==7.4.2
pillow==10.1.0
APScheduler==3.10.4
python-dateutil==2.8.2
